requirements.txt
beautifulsoup4==4.9.3
dnspython==1.16.0
openpyxl==3.0.5
pymongo==3.11.3
python-telegram-bot==13.1
requests==2.25.0
urllib3==1.26.2
yapf==0.30.0
